backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
google-genai==1.60.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
python-multipart==0.0.22
opik==1.9.98
googlemaps>=4.10.0  # Google Maps APIs (Places, Geocoding)

pillow>=10.0.0
python-magic-bin>=0.4.14  # Windows compatible

sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

email-validator==2.3.0